Module ncomm::prelude

Re-export of NComm traits to make working with NComm a bit easier

Modules

client_server

NComm Client Server Traits.

executor

An executor handles the scheduling and execution of nodes.

node

A Singular Unit of Work.

publisher_subscriber

Publisher -> Subscriber Communication

update_client_server

NComm Update Server and Client.

Enums

ExecutorState

The current state an executor is in.

Traits

Client

A common abstraction for all NComm clients to allow for the creation of a common method of sending requests and receiving responses.

Executor

An executor handles the scheduling and execution of nodes

Node

A Node represents a singular process that performs some singular purpose

Publisher

The basic publisher trait that enables the publishing of data to some endpoint for subscribers to read.

Server

A common abstraction for all NComm servers that outlines the necessary base requirements for all NComm servers

Subscriber

The basic subscriber trait that enables for the reading of data from some endpoint.

UpdateClient

A common abstraction for all NComm update clients

UpdateServer

A common abstraction for all NComm update servers