Trait ncomm::prelude::Subscriber

pub trait Subscriber {
    type Target;

    // Required method
    fn get(&mut self) -> &Self::Target;
}

The basic subscriber trait that enables for the reading of data from some endpoint.

Note: For convenience sake, Subscriber is actually just an alias for the Deref trait

Required Associated Types

type Target

The type of data stored in the subscriber

Required Methods

fn get(&mut self) -> &Self::Target

Update the current data in the subscriber and return a reference to the current data