Trait ncomm::prelude::update_client_server::UpdateClient

pub trait UpdateClient {
    type Request;
    type Update;
    type Response;
    type Error;

    // Required methods
    fn send_request(
        &mut self,
        request: Self::Request,
    ) -> Result<(), Self::Error>;
    fn poll_for_update(
        &mut self,
    ) -> Result<Option<(Self::Request, Self::Update)>, Self::Error>;
    fn poll_for_updates(
        &mut self,
    ) -> Vec<Result<(Self::Request, Self::Update), Self::Error>>;
    fn poll_for_response(
        &mut self,
    ) -> Result<Option<(Self::Request, Self::Response)>, Self::Error>;
    fn poll_for_responses(
        &mut self,
    ) -> Vec<Result<(Self::Request, Self::Response), Self::Error>>;
}

A common abstraction for all NComm update clients

Required Associated Types

type Request

The type of data used as a request by the client

type Update

The type of data used as an update from the server

type Response

The type of data used as a response from the server

type Error

The type of error from sending or receiving data from the update server

Required Methods

fn send_request(&mut self, request: Self::Request) -> Result<(), Self::Error>

Send a request to the server this client is associated with

fn poll_for_update( &mut self, ) -> Result<Option<(Self::Request, Self::Update)>, Self::Error>

Poll for a singular update from the server

fn poll_for_updates( &mut self, ) -> Vec<Result<(Self::Request, Self::Update), Self::Error>>

Poll for updates from the server

fn poll_for_response( &mut self, ) -> Result<Option<(Self::Request, Self::Response)>, Self::Error>

Poll for a singular response from the server

fn poll_for_responses( &mut self, ) -> Vec<Result<(Self::Request, Self::Response), Self::Error>>

Poll for responses from the server